+++ gitoxide-core/src/repository/mod.rs
@@ -4,7 +4,7 @@ use anyhow::{Context as AnyhowContext, Result};
 use git_repository as git;
 
 pub fn init(directory: Option<PathBuf>) -> Result<git::discover::repository::Path> {
-    git_repository::path::create::into(directory.unwrap_or_default(), git_repository::Kind::WorkTree)
+    git_repository::create::into(directory.unwrap_or_default(), git_repository::Kind::WorkTree)
         .with_context(|| "Repository initialization failed")
 }
